add login requird

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import Product, Category
 from django.db.models import Q
 from django.db.models.functions import Lower
@@ -72,11 +73,17 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
-
+@login_required
 def add_product(request):
     """
     Allow staff or superusers to add a new product to the store.
     """
+    if not request.user.is_superuser:
+        messages.error(
+            request, "Only admin can add products"
+        )
+        return redirect(reverse("home"))
+    
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -90,10 +97,17 @@ def add_product(request):
         request, 'products/add_product.html', {
             'form': form, 'products': products})
 
+@login_required
 def edit_product(request, product_id):
     """
     Allow staff or superusers to edit an existing product's details.
     """
+    if not request.user.is_superuser:
+        messages.error(
+            request, "Only admin can add products"
+        )
+        return redirect(reverse("home"))
+    
     product = get_object_or_404(Product, id=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -109,10 +123,17 @@ def edit_product(request, product_id):
     }
     return render(request, 'products/edit_product.html', context)
 
+@login_required
 def delete_product(request, product_id):
     """
     Allow staff or superusers to delete a product from the store.
     """
+    if not request.user.is_superuser:
+        messages.error(
+            request, "Only admin can add products"
+        )
+        return redirect(reverse("home"))
+    
     product = get_object_or_404(Product, id=product_id)
     product.delete()
     messages.success(request, 'Product deleted successfully!')
